Finally figured out how to zoom on the mouse pos.
The code is dirty, and needs cleaning. But the implementation
is simple. Turns out QGraphicsView doesn't use the translation
values of the transformation that can be set with setTransform,
only rotation and scale. centerOn must be used to scroll the
view to the desired place.
Current code doesn't handle rotation and if the mouse is
outside the view when zooming in or out (as when the toolbar
zoom buttons are clicked), the last recorded mouse position
inside the view is used as anchor, which is weird.