Working on improving the disk performance.
Instead of waking up the fetcher thread on each fetchTile, I let
TileLayer decide when to call it. That gives it the chance of
queueing many fetchTile requests and then waking only once to
handle them all.
I'm also testing whether a thread for disk reads is really necessary,
given that TileFetcher runs in its own thread. Not having the disk
reads in a different thread will also help meet the goal of reading
the mgm files more efficiently, i.e., not opening and closing for
each tile.
Right now, the code is reading and emitting all tiles it finds in
the mgm, but that seems to be overkill. By examining the diskRequests
queue in TileFetcher, the code can be smarter about that.