Moved wakeUp back to fetchTile and removed the call from TileLayer
The thing is that, sometimes, maybe for some race condition, but
I'm still not sure, some fetchTiles signals were processed after
the wakeUp in TileLayer and they stalled there, with no one to
wake the fetcher up to handle them.
I assumed that, even with queued connections between threads, the
order of processing signals would match the order of their emission.
But it's better not to rely on that.