TileFetcher now does the disk reading.
And it uses its knowledge of the diskRequestQueue to effectively
reduce the amount of mgm opening, reading and closing. Now, after
opening one mgm, all request in the queue that can be served with
that mgm are served, and all requests that belong to that mgm but
are not found in the mgm are delegated to the network queue.