Don't remove extra mgm absent tiles from the disk queue.
I was checking the diskRequests queue for tiles that belonged
to the current mgm file, but were not present in the mgm. I then
immediately removed them from the disk queue and added them to
the network queue. The problem with that is, when using single
file tiles, a given tile (x,y) may indeed be present in the mgm,
while its neighbour (x+1,y) is not. However, it may happen that,
although (x+1,y) is not in the mgm, it is indeed available as a
single file, but it is waiting its turn in the queue. Removing
it from the queue is then premature. On the other hand, removing
the (x,y) tile, which is the one explicitly requested, is, of
course, required.