Flicker was not because centerScene was called in paint
Quick movements flickered because centerScene was just subtracting
a projectionWidth each time. So, in the paint event, sometimes
that subtraction wasn't enough to bring it to the "home" world,
and the next paint event, scheduled by centerOn, moved the world
again, and so on until the home world was reached and painting
finished.